Avoid using the same temporary launch file when triggering multiple debug sessions in parallel
